backend/requirements.txt
# FastAPI Backend
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic==2.9.0
httpx==0.27.0

# Exa Search
exa-py==1.4.0

# xAI SDK
xai-sdk>=0.1.0

# mem0 local dependencies
openai>=1.0.0
qdrant-client==1.12.0

# Vector embeddings (for mem0 local mode)
sentence-transformers==3.0.0

# SuperMemory abstraction (local SQLite)
aiosqlite==0.20.0

# Utilities
numpy>=1.26.0
